Extract SpeculativeConfig from configs
